requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
pyDolarVenezuela==1.4.2
python-dotenv==1.0.1